app/backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0
pydantic>=2.5.0
pydantic-settings>=2.0.0

# Conversion libraries
rlottie-python==1.3.6
pyoxipng>=9.1.1
Pillow>=10.0.0
python-telegram-bot>=21.0
aiofiles>=23.0.0
